refactor: restructure forms to improved organization
