--- src/components/RegistrationForm/InputRegister.tsx
@@ -1,130 +1,0 @@
-import { Visibility, VisibilityOff } from "@mui/icons-material";
-import {
-  Box,
-  Button,
-  IconButton,
-  InputAdornment,
-  TextField,
-} from "@mui/material";
-import { useState } from "react";
-
-export const InputRegister = () => {
-  const [firstName, setFirstName] = useState("");
-  const [lastName, setLastName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [showPassword, setShowPassword] = useState(false);
-
-  //TO-DO falta implementar toast quando back estiver ok
-
-  const handleLastNameChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setLastName(event.target.value);
-  };
-
-  const handleEmailChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setEmail(event.target.value);
-  };
-
-  const handlePasswordChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setPassword(event.target.value);
-  };
-  const handleClickShowPassword = () => {
-    setShowPassword(!showPassword);
-  };
-
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    console.log(firstName, lastName, email, password);
-  };
-
-  const handleMouseDownPassword = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    event.preventDefault();
-  };
-
-  const handleFirstNameChange = (
-    event: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    setFirstName(event.target.value);
-  };
-
-  return (
-    <Box component="form" onSubmit={handleSubmit} sx={{ mt: 1, width: "100%" }}>
-      <Box
-        display={"flex"}
-        width={"100%"}
-        sx={{
-          gap: 1,
-        }}
-      >
-        <TextField
-          margin="normal"
-          required
-          fullWidth
-          id="firstName"
-          label="Nome"
-          name="firstName"
-          autoComplete="given-name"
-          value={firstName}
-          onChange={handleFirstNameChange}
-        />
-
-        <TextField
-          margin="normal"
-          required
-          fullWidth
-          id="lastName"
-          label="Sobrenome"
-          name="lastName"
-          autoComplete="family-name"
-          value={lastName}
-          onChange={handleLastNameChange}
-        />
-      </Box>
-
-      <TextField
-        margin="normal"
-        required
-        fullWidth
-        id="email"
-        label="Email address"
-        name="email"
-        autoComplete="email"
-        value={email}
-        onChange={handleEmailChange}
-      />
-
-      <TextField
-        margin="normal"
-        required
-        fullWidth
-        name="password"
-        label="Password"
-        type={showPassword ? "text" : "password"}
-        id="password"
-        autoComplete="current-password"
-        value={password}
-        onChange={handlePasswordChange}
-        InputProps={{
-          endAdornment: (
-            <InputAdornment position="end">
-              <IconButton
-                aria-label="toggle password visibility"
-                onClick={handleClickShowPassword}
-                onMouseDown={handleMouseDownPassword}
-                edge="end"
-              >
-                {showPassword ? <VisibilityOff /> : <Visibility />}
-              </IconButton>
-            </InputAdornment>
-          ),
-        }}
-      />
-
-      <Button type="submit" fullWidth variant="contained" sx={{ mt: 3, mb: 2 }}>
-        CADASTRAR
-      </Button>
-    </Box>
-  );
-};
