+++ src/routes.tsx
@@ -1,12 +1,12 @@
-import { createBrowserRouter } from "react-router-dom";
-import Layout from "./Layout";
-import HomePage from "./pages/Home";
-import LoginPage from "./pages/LoginPage";
-import { RegisterPage } from "./pages/RegisterPage";
+import { createBrowserRouter } from 'react-router-dom';
+import Layout from './Layout';
+import HomePage from './pages/Home';
+import LoginPage from './pages/LoginPage';
+import RegisterPage from './pages/RegisterPage';
 
 const router = createBrowserRouter([
   {
-    path: "/",
+    path: '/',
     element: <Layout />,
     children: [
       { index: true, element: <HomePage /> },
@@ -14,11 +14,11 @@ const router = createBrowserRouter([
     ],
   },
   {
-    path: "/login",
+    path: '/login',
     element: <LoginPage />,
   },
   {
-    path: "/registrar",
+    path: '/cadastro',
     element: <RegisterPage />,
   },
 ]);
